Restore size_t

+++ src/mode/CBC.c
@@ -1,6 +1,6 @@
 #include "CBC.h"
 
-void CBC_enc(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *IV, uint8_t *key, int block_size, size_t len, unsigned char *src, unsigned char *des) {
+void CBC_enc(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *IV, uint8_t *key, size_t block_size, size_t len, unsigned char *src, unsigned char *des) {
 
     uint8_t key2[16];
 	
@@ -24,7 +24,7 @@ void CBC_enc(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *IV, uint
     }
 }
 
-void CBC_dec(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *IV, uint8_t *key, int block_size, size_t len, unsigned char *src, unsigned char *des) {
+void CBC_dec(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *IV, uint8_t *key, size_t block_size, size_t len, unsigned char *src, unsigned char *des) {
 
     uint8_t key2[16];
