+++ src/mode/ECB.c
@@ -1,7 +1,7 @@
 #include "ECB.h"
 
 //ECB운용모드
-void ECB(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *key, int block_size, size_t len, unsigned char *src, unsigned char *des) {
+void ECB(void (*cipher)(uint8_t *, uint8_t *, uint8_t *) , uint8_t *key, size_t block_size, size_t len, unsigned char *src, unsigned char *des) {
 
     uint8_t key2[16];
